// src/types/bill.ts
export interface ApiBill {
  SearchScore: number;
  BillId: number;
  BillNumberPrefix: string | null;
  BillNumber: number;
  BillNumberSuffix: string | null;
  BillNumberFormatted: string;
  LongTitleEn: string;
  LongTitleFr: string;
  ShortTitleEn: string;
  ShortTitleFr: string;
  LatestBillMilestoneId: number;
  LatestCompletedMajorStageEn: string;
  LatestCompletedMajorStageFr: string;
  LatestCompletedMajorStageChamberId: number;
  DidReinstateFromPreviousSession: boolean;
  PassedHouseFirstReadingDateTime: string | null;
  PassedHouseSecondReadingDateTime: string | null;
  PassedHouseThirdReadingDateTime: string | null;
  PassedSenateFirstReadingDateTime: string | null;
  PassedSenateSecondReadingDateTime: string | null;
  PassedSenateThirdReadingDateTime: string | null;
  ReceivedRoyalAssentDateTime: string | null;
  ParlSessionCode: string;
  ParlSessionEn: string;
  ParlSessionFr: string;
  ParliamentNumber: number;
  SessionNumber: number;
  OriginatingChamberId: number;
  BillTypeId: number;
  BillTypeEn: string;
  BillTypeFr: string;
  CurrentStatusId: number;
  CurrentStatusEn: string;
  CurrentStatusFr: string;
  MinistryId: number;
  SponsorId: number;
  SponsorEn: string;
  SponsorFr: string;
  PoliticalAffiliationId: number;
  IsFromCurrentSession: boolean;
  Highlight: string | null;
  LatestBillTextTypeId: number;
  LatestActivityEn: string;
  LatestActivityFr: string;
  LatestActivityDateTime: string;
} 